refactor(permissions): co-locate hooks with their classes
PERMISSIONS and PERMISSION_SCOPES were defined in a standalone app.core.permissions.hooks
module typed as SingleNamedItemHook[HasName] purely to avoid a circular import with the
Permission / PermissionScope classes. Move each hook into the module of the class it
stores (PERMISSIONS into app.core.permissions, PERMISSION_SCOPES into
app.core.permissions.scopes) and type it with the concrete class; delete the now-empty
hooks module.

HasName disappears from the permissions layer; it survives only as the TypeVar bound of
the generic SingleNamedItemHook in app/lib/hooks.py, where it lets the hook read .name.
This also drops a now-redundant cast(PermissionScope, …) in the registry — iter_values()
returns the concrete type directly.
