+++ app/lib/permissions/__init__.py
@@ -16,10 +16,11 @@ from app.core.permissions import (
     EMAIL_WHITELIST_CREATE,
     EMAIL_WHITELIST_DELETE,
     EMAIL_WHITELIST_READ,
+    PERMISSIONS,
     Permission,
 )
 from app.core.permissions.exceptions import PermissionScopeNotFound, RoleNotFound
-from app.core.permissions.hooks import PERMISSION_SCOPES, PERMISSIONS
+from app.core.permissions.scopes import PERMISSION_SCOPES
 from app.core.permissions.utils import assign_role, can, has_role, revoke_role
 
 __all__ = [
